IdentityBadge: consolidate popover title and fix network selection
